feat(mundo): ids de bloco em 16 bits, com conversão automática dos saves antigos
O chunk era 1 byte por célula: teto de 256 ids, com só 5 livres depois
da cabeceira da cama — os blocos de circuito lógico não cabiam. Agora o
chunk é Uint16Array na memória, e bloco pode ir até 899 (os itens
começam em 900; há portão em teste).

No disco e no fio cada chunk leva 1 byte de largura: estreito (u8) quando
todos os ids dele cabem em 255, largo (u16) só quando não cabem. Mundo
que existe hoje cresce 1 byte por chunk — o Wi-Fi da escola e os .ljw
distribuídos não pagam o dobro. Formatos novos LJW1 (snapshot), LJC1
(streaming) e LJS3 (save esparso); LJW0, LJC0 e LJS2 continuam sendo
lidos pra sempre, e o save lido deles vem marcado `legado`.

Conversão automática, sem passo manual: ler o antigo é permanente e
regravar já escreve o novo. Antes da 1ª gravação por cima, o original é
guardado — no host, `<nome>.antes-ids16.ljw` na pasta do mundo (no boot
converte na hora); no singleplayer, `dataAntesIds16` no registro do
IndexedDB. Modelos de cenarios/ e aulas só leitura nunca são escritos.

Pegadinhas cobertas: `Uint8Array.set(Uint16Array)` compila e corta id
≥ 256 em silêncio — a vizinhança do mesher (o que vai pro Worker) virou
Uint16Array, com teste. E o `ehEstreito` célula a célula deixou o
encodeSave 8× mais lento, e o save do Ctrl+C passou a perder a corrida
com o npx/tsx que embrulha o host (bug-666); varrendo 2 células por vez
ele volta a ~6 ms.

Provas: hash dos blocos de cenarios/aula1 e server/world.ljw idêntico ao
medido com o código de 1 byte; 18 testes novos (1014 verdes); smoke 16/16;
no cliente real a loja (mundo denso) e um mundo E por streaming; save
antigo real convertido no host (cópia idêntica, 2ª subida sem converter)
e no IndexedDB de um navegador real.

+++ client/src/chunks.ts
@@ -449,7 +449,7 @@ export class ChunkRenderer {
         this.caminho = c.carga ? "carga" : "fila";
         const t0 = performance.now();
         const versao = this.novaVersao(key);
-        let viz: Uint8Array | null;
+        let viz: Uint16Array | null;
         try {
           viz = extrairVizinhanca(this.world, c.cx, c.cy, c.cz);
         } catch (e) {
